+++ stress/src/random.rs
@@ -25,5 +25,12 @@ fn main() {
 }
 
 fn test_random_generation() {
-    let _i1 = CURRENT_RNG.with_borrow_mut(|rng| [rng.gen_range(0..10), rng.gen_range(0..10), rng.gen_range(0..10)]);
+    let _i1 = CURRENT_RNG.with(|rng| {
+        let mut rng = rng.borrow_mut();
+        [
+            rng.gen_range(0..10),
+            rng.gen_range(0..10),
+            rng.gen_range(0..10),
+        ]
+    });
 }